backend/requirements.txt
flask
flask-cors
python-dotenv
werkzeug
replicate
supabase
pillow
requests